compute omi stats

+++ CMR_queries/cme_stats.py
@@ -5,13 +5,21 @@
 import json
 import re
 from collections import defaultdict
+from enum import Enum
+import os
+
+
+class CMRSearchType(Enum):
+    SCIENCE_KEYWORD = 0,
+    KEYWORD = 1,
+    BOTH = 2
+
 
 with open('3-22-15-Aura_omi_features.json', encoding='utf-8') as f:
     features = json.load(f)
 
 with open('20-20-16_omi_papers_key_title_ground_truth.json', encoding='utf-8') as f:
     key_title_ground_truth = json.load(f)
-
 
 
 def format_lot(lot):
@@ -30,7 +38,7 @@ def correct_missed_extraneous(ground_truths, predictions):
     return correct, missed, extraneous
 
 
-def dump_data(key, features, csv, manually_reviewed=None, title='', running_cme_stats=None, n=1):
+def dump_data(key, features, csv, manually_reviewed=None, title='', running_cme_stats=None, n=1, dataset_search_type=None):
     summary_stats = features['summary_stats']
     couples = sorted(list(summary_stats['valid_couples'].items()), key=lambda x: x[1], reverse=True)
     models = sorted(list(summary_stats['models'].items()), key=lambda x: x[1], reverse=True)
@@ -45,8 +53,36 @@ def dump_data(key, features, csv, manually_reviewed=None, title='', running_cme_
     # get SINGLE TOP CMR results
     cmr_results = set()
     for inner_key, inner_value in features['cmr_results']['pairs'].items():
-        datasets = inner_value['science_keyword_search']['dataset']
-        # datasets = inner_value['keyword_search']['dataset']
+
+        if dataset_search_type == CMRSearchType.SCIENCE_KEYWORD:
+            datasets = inner_value['science_keyword_search']['dataset']
+        elif dataset_search_type == CMRSearchType.KEYWORD:
+            datasets = inner_value['keyword_search']['dataset']
+        elif dataset_search_type == CMRSearchType.BOTH:
+            # merge the two lists together, alternating order
+            l1 = inner_value['science_keyword_search']['dataset']
+            l2 = inner_value['keyword_search']['dataset']
+
+            i, j, datasets_temp = 0, 0, []
+            while i < len(l1) and j < len(l2):
+                datasets_temp.append(l1[i])
+                datasets_temp.append(l2[j])
+                i += 1
+                j += 1
+            if i < len(l1):
+                datasets_temp += l1[i:]
+            elif j < len(l2):
+                datasets_temp += l2[j:]
+
+            # remove duplicates
+            seen = set()
+            datasets = []
+            for i in range(len(datasets_temp)):
+                if datasets_temp[i] in seen:
+                    continue
+                seen.add(datasets_temp[i])
+                datasets.append(datasets_temp[i])
+
         if len(datasets) >= 1:
             for predic in datasets[:n]:
                 cmr_results.add(predic)
@@ -71,9 +107,17 @@ def dump_data(key, features, csv, manually_reviewed=None, title='', running_cme_
 
 if __name__ == '__main__':
     n = 1
+    cmr_search_type = CMRSearchType.BOTH
+    sub_folder = f'omi_{cmr_search_type.name.lower()}/'
+    base_location = 'stats/' + sub_folder
 
-    while n <= 20:
-        filename = f'Aura_omi_cme_top_{n}'
+    correct, missed, extraneous = [], [], []
+
+    if not os.path.exists(base_location):
+        os.makedirs(base_location)
+
+    while n <= 9:
+        filename = base_location+ f'Aura_omi_cme_top_{n}_{cmr_search_type.name.lower()}'
         added_pdfs = set()
         running_cme_stats = {
             "correct_count": 0,
@@ -90,15 +134,19 @@ if __name__ == '__main__':
             added_pdfs.add(pdf_key)
             if pdf_key in features:
                 csv = dump_data(pdf_key, features[pdf_key], csv, manually_reviewed=value, title=value['title'], running_cme_stats=running_cme_stats,
-                                n=n)
+                                n=n, dataset_search_type=cmr_search_type)
 
         for key, value in features.items():
             if key not in added_pdfs:
-                csv = dump_data(key, value, csv)
+                csv = dump_data(key, value, csv, dataset_search_type=cmr_search_type)
 
         running_cme_stats['correct_dict'] = dict(sorted(running_cme_stats['correct_dict'].items(), key=lambda x: x[1], reverse=True))
         running_cme_stats['missed_dict'] = dict(sorted(running_cme_stats['missed_dict'].items(), key=lambda x: x[1], reverse=True))
         running_cme_stats['extraneous_dict'] = dict(sorted(running_cme_stats['extraneous_dict'].items(), key=lambda x: x[1], reverse=True))
+
+        if os.path.exists(filename + '.json'):
+            print("\n\nFile with name already exists\n\n")
+            exit()
 
         with open(filename + '.json', 'w', encoding='utf-8') as f:
             json.dump(running_cme_stats, f, indent=4)
@@ -106,4 +154,19 @@ if __name__ == '__main__':
         with open(filename + '.csv', 'w', encoding='utf-8') as f:
             f.write(csv)
 
+        # save the counts for correct, missed, extraneous into the local arrays
+        correct.append(running_cme_stats['correct_count'])
+        missed.append(running_cme_stats['missed_count'])
+        extraneous.append(running_cme_stats['extraneous_count'])
+
         n += 1
+
+    # save a file with the three lists for correct missed and extraneous
+    summary_dict = {
+        "cmr_mode": cmr_search_type.name.lower(),
+        "correct_counts": correct,
+        "missed_counts": missed,
+        "extraneous_counts": extraneous,
+    }
+    with open(base_location + f'{cmr_search_type.name.lower()}_summary_counts.json', 'w', encoding='utf-8') as f:
+        json.dump(summary_dict, f)
